student_revinskaya_ekaterina/src/main/java/lv/javaguru/travel/insurance/core/validations/TravelRequestSelectedRisksValidation.java
package lv.javaguru.travel.insurance.core.validations;

import lv.javaguru.travel.insurance.dto.TravelCalculatePremiumRequest;
import lv.javaguru.travel.insurance.dto.ValidationError;
import org.springframework.stereotype.Component;

import java.util.Optional;
@Component
class TravelRequestSelectedRisksValidation implements TravelRequestValidation {
    @Override
    public Optional<ValidationError> validate(TravelCalculatePremiumRequest request) {
        return (request.getSelected_risks() == null || request.getSelected_risks().isEmpty())
                ? Optional.of(new ValidationError("selected_risks", "Must not be empty!"))
                : Optional.empty();
    }
}
